Refactor FS traits
