+++ marlin/src/fiat_shamir/traits/fiat_shamir.rs
@@ -15,12 +15,7 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use snarkvm_fields::{PrimeField, ToConstraintField};
-use snarkvm_gadgets::{
-    fields::FpGadget,
-    utilities::{boolean::Boolean, uint::UInt8},
-};
-use snarkvm_nonnative::{params::OptimizationType, NonNativeFieldVar};
-use snarkvm_r1cs::{ConstraintSystem, SynthesisError};
+use snarkvm_nonnative::params::OptimizationType;
 
 use rand_core::RngCore;
 
@@ -42,71 +37,4 @@ pub trait FiatShamirRng<TargetField: PrimeField, BaseField: PrimeField>: RngCore
     fn squeeze_native_field_elements(&mut self, num: usize) -> Vec<BaseField>;
     /// Takes out field elements of 128 bits.
     fn squeeze_128_bits_nonnative_field_elements(&mut self, num: usize) -> Vec<TargetField>;
-}
-
-/// Vars for a RNG for use in a Fiat-Shamir transform.
-pub trait FiatShamirRngVar<TargetField: PrimeField, BaseField: PrimeField, PFS: FiatShamirRng<TargetField, BaseField>>:
-    Clone
-{
-    /// Create a new RNG.
-    fn new<CS: ConstraintSystem<BaseField>>(cs: CS) -> Self;
-
-    /// Instantiate from a plaintext fs_rng.
-    fn constant<CS: ConstraintSystem<BaseField>>(cs: CS, pfs: &PFS) -> Self;
-
-    /// Take in field elements.
-    fn absorb_nonnative_field_elements<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        elems: &[NonNativeFieldVar<TargetField, BaseField>],
-        ty: OptimizationType,
-    ) -> Result<(), SynthesisError>;
-
-    /// Take in field elements.
-    fn absorb_native_field_elements<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        elems: &[FpGadget<BaseField>],
-    ) -> Result<(), SynthesisError>;
-
-    /// Take in bytes.
-    fn absorb_bytes<CS: ConstraintSystem<BaseField>>(&mut self, cs: CS, elems: &[UInt8]) -> Result<(), SynthesisError>;
-
-    /// Output field elements.
-    fn squeeze_native_field_elements<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        num: usize,
-    ) -> Result<Vec<FpGadget<BaseField>>, SynthesisError>;
-
-    /// Output field elements.
-    fn squeeze_field_elements<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        num: usize,
-    ) -> Result<Vec<NonNativeFieldVar<TargetField, BaseField>>, SynthesisError>;
-
-    /// Output field elements and the corresponding bits (this can reduce repeated computation).
-    #[allow(clippy::type_complexity)]
-    fn squeeze_field_elements_and_bits<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        num: usize,
-    ) -> Result<(Vec<NonNativeFieldVar<TargetField, BaseField>>, Vec<Vec<Boolean>>), SynthesisError>;
-
-    /// Output field elements with only 128 bits.
-    fn squeeze_128_bits_field_elements<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        num: usize,
-    ) -> Result<Vec<NonNativeFieldVar<TargetField, BaseField>>, SynthesisError>;
-
-    /// Output field elements with only 128 bits, and the corresponding bits (this can reduce
-    /// repeated computation).
-    #[allow(clippy::type_complexity)]
-    fn squeeze_128_bits_field_elements_and_bits<CS: ConstraintSystem<BaseField>>(
-        &mut self,
-        cs: CS,
-        num: usize,
-    ) -> Result<(Vec<NonNativeFieldVar<TargetField, BaseField>>, Vec<Vec<Boolean>>), SynthesisError>;
 }
